MENTOR Greg:  HW REORG: FIX 2022-12-05/
